Adds checkstyle ClassTypeParameterName (#1004)
* Uncomments checkstyle

* Removes GoogleSheetsTask
